Global reduction works! And it is beautiful. But...
Notice how many non-interacting nodes there are midway through the
reduction, in comparison to the number of interacting nodes. If only a
small part of nodes are active at any point, then this kind of strategy
will not work very well. In an hypothetical machine where every core is
active simultaneously, global movement passes would be free. But in a
GPU, this is costly.

The good thing about this implementation is that is successfully reduces
up to 2048 inside a streaming multiprocessor in a quite fast way. If a
high % of the graph was always active, then it would be performing
absurd amount of parallel reductions. Look at this, for example:

https://imgur.com/a/9t84lXT

This is the third example on the code. Each yellow bar represents an
inactive node, each red bar represents a node active in a duplication,
and each green bar represents a node active in an annihilation. A good %
of the graph is active. In this case, this approach might be fast.

https://imgur.com/a/VKuqven

This is the second example. The number of active nodes is minimal
compared to the number of inactive ones. In this case, the global
movement and allocation work will probably far outweight the speedups of
local SM reductions.  In those cases, I guess the only thing that
matters is the memory throughput, simple as that. Even CPU/GPU wouldn't
make a huge difference since the bottleneck of the algorithm is not
computational, but simply and plainly how fast it reads memory and
writes it back.

An ideal architecture would perhaps be just a really, really huge global
memory, which enabled tons and tons of parallel reads, connected to
millions of ultra-thin processors?